codex(web): extract history idempotency store
